+++ Libplanet.Net/Transports/DifferentAppProtocolVersionException.cs
@@ -7,7 +7,7 @@ namespace Libplanet.Net.Transports
 {
     /// <summary>
     /// The exception that is thrown when the version of the
-    /// <see cref="Message"/> that <see cref="Swarm{T}"/> received
+    /// <see cref="Message"/> that <see cref="Swarm"/> received
     /// is different.
     /// </summary>
     [Serializable]
@@ -19,9 +19,9 @@ namespace Libplanet.Net.Transports
         /// </summary>
         /// <param name="message">Specifies an <see cref="Exception.Message"/>.</param>
         /// <param name="expectedAppProtocolVersion">The protocol version of
-        /// the local <see cref="Swarm{T}"/>.</param>
+        /// the local <see cref="Swarm"/>.</param>
         /// <param name="actualAppProtocolVersion">The protocol version of the
-        /// <see cref="BoundPeer"/> that the local <see cref="Swarm{T}"/> is trying to connect
+        /// <see cref="BoundPeer"/> that the local <see cref="Swarm"/> is trying to connect
         /// to.</param>
         /// <param name="trusted">Whether <paramref name="actualAppProtocolVersion"/>
         /// is signed by a trusted signer.</param>
@@ -50,13 +50,13 @@ namespace Libplanet.Net.Transports
         }
 
         /// <summary>
-        /// The protocol version of the current <see cref="Swarm{T}"/>.
+        /// The protocol version of the current <see cref="Swarm"/>.
         /// </summary>
         public AppProtocolVersion ExpectedApv { get; }
 
         /// <summary>
         /// The protocol version of the <see cref="BoundPeer"/> that the
-        /// <see cref="Swarm{T}" /> is trying to connect to.
+        /// <see cref="Swarm" /> is trying to connect to.
         /// </summary>
         public AppProtocolVersion ActualApv { get; }